requirements.txt
ollama
numpy
pandas
scikit-learn
faiss-cpu
